[#17] refactor : fetch file at assembler
