+++ src/components/assembler/AssembleFilePanel.tsx
@@ -1,12 +1,11 @@
-import { useRecoilState, useRecoilValue } from "recoil";
+import { useRecoilValue } from "recoil";
 import {
   selectedAssemblyFileState,
   selectedFileContentState,
 } from "../../recoil/state";
 import Panel from "../common/Panel";
-import { useEffect } from "react";
-import { HL_ORANGE } from "../../styles/color";
 import TopTab from "../common/TopTab";
+import { HL_ORANGE } from "../../styles/color";
 
 const AssembleFilePanel = ({
   highlightNumbers,
@@ -14,28 +13,14 @@ const AssembleFilePanel = ({
   highlightNumbers: number[];
 }) => {
   const selectedAssemblyFile = useRecoilValue(selectedAssemblyFileState);
-  const [fileContent, setFileContent] = useRecoilState(
-    selectedFileContentState
-  );
-
-  useEffect(() => {
-    const fetchFile = async (filePath: string) => {
-      await fetch(filePath)
-        .then((response) => response.text())
-        .then((text) => {
-          setFileContent(text.split("\n"));
-        });
-    };
-    const filePath = `sample_input/${selectedAssemblyFile}`;
-    fetchFile(filePath);
-  }, [selectedAssemblyFile, setFileContent]);
+  const fileContent = useRecoilValue(selectedFileContentState);
 
   return (
     <div style={{ flexDirection: "row" }}>
       <TopTab title={selectedAssemblyFile} isBinary={false} />
       <Panel
         highlightNumbers={highlightNumbers}
-        data={fileContent ? fileContent : []}
+        data={fileContent || []}
         highlightColor={HL_ORANGE}
         width={"592px"}
         type={"assembly"}
